handled max biginteger

+++ Quiz/Classes/Question.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -8,11 +9,11 @@ namespace Quiz.Classes
 {
     internal class Question
     {
-        public float CorrectAnswer { get; set; }
+        public BigInteger CorrectAnswer { get; set; }
         public Equation Equation { get; set; }
-        public float[] Answers { get; set; }
+        public BigInteger[] Answers { get; set; }
 
-        public Question(Equation equation, float correctAnswer, float[] answers)
+        public Question(Equation equation, BigInteger correctAnswer, BigInteger[] answers)
         {
             Equation = equation;
             CorrectAnswer = correctAnswer;
